Divkit пока без темы и без навигации назад

+++ feature/todo/src/main/java/com/cleverpumpkin/todo/presentation/screens/todo_detail_screen/TodoDetailViewModel.kt
@@ -94,7 +94,9 @@ class TodoDetailViewModel @Inject constructor(
     }
 
     fun deleteItem() = viewModelScope.launch {
-        processResponse(repository.deleteTodoItem(createTodo()))
+        if (_uiState.value.id.isNotEmpty()) {
+            processResponse(repository.deleteTodoItem(createTodo()))
+        }
     }
 
     private fun processResponse(response: Response<*>, onSuccessAction: (() -> Unit)? = null) {
